Added Pixel FitnessTracker App
The Pixel Fitness Tracker app is an app made using Room, Coroutines, Lifecycle Awareness and Fragment Navigation to display a RecyclerView that contains a list of workouts. The app allows you to delete workouts from the list, update the calm weight and max weight of the workout and allows you to add new workouts. Data is stored in a database for data persistence.